% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbGaPdb-class}
\alias{dbGaPdb-class}
\title{Interact with dbGaPdb}
\description{
Interact with dbGaPdb
}
\section{Methods}{

\describe{
\item{\code{connect(path)}}{Connect to the database
This is a test}
}}

