% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Basic hello, world example.}
\usage{
hello(phrase = "World")
}
\arguments{
\item{phrase}{The phrase to replace "World" with?}
}
\description{
In terms of a more detailed explanation,
this function does some really cool hello-saying.
}
\examples{
hello()
hello('world')

}
