% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dbGapdb_sqlite.R
\name{pull_dbGaPdb_sqlite}
\alias{pull_dbGaPdb_sqlite}
\title{Downloads dbGaPdb sqlite from temp link and unzips it}
\usage{
pull_dbGaPdb_sqlite(local_path = tempdir(),
  sqlite_db_name = "dbGaPdb.sqlite",
  link = "https://helix.nih.gov/~mcgaugheyd/dbGapdb.sqlite.gz")
}
\description{
Downloads dbGaPdb sqlite from temp link and unzips it
}
